requirements.txt
absl-py==0.9.0
allure-pytest==2.8.36
allure-python-commons==2.8.36
aniso8601==8.0.0
apipkg==1.5
asgiref==3.2.3
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
execnet==1.8.0
Faker==6.6.1
idna==2.10
iniconfig==1.1.1
packaging==20.9
pluggy==0.13.1
py==1.10.0
pyparsing==2.4.7
pytest==6.2.2
pytest-forked==1.3.0
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.1
python-dateutil==2.8.1
requests==2.25.1
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
